Created NbaService, FakeNbaService classes, made logger a parameter.
I initially started this because I wanted to make nba_data.py use the right logger when it was used by the game_thread_bot. It seemed like the simplest way to do that was to make the logger a configurable parameter but this lead me down a path to making all the code that calls the NBA APIs into an injectable class.

This change allows unit tests to fake or mock NbaService class instead of trying to mock requests and it allows the logger parameter to be configured differently in tests, which creates a lot less noise when running tests and it allows the game_thread_bot and sidebar bot to log their messages independently.